scanner-service/requirements.txt
# GitScan Security Scanner - Python Dependencies

# Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Async/Background Tasks
celery==5.3.4
redis==5.0.1

# Security Scanning
bandit==1.7.6          # Python security linter
semgrep==1.52.0        # Multi-language static analysis
safety==2.3.5          # Dependency vulnerability scanner

# Code Parsing
tree-sitter==0.20.4
libcst==1.1.0          # Python AST

# GitHub Integration
PyGithub==2.1.1
gitpython==3.1.40

# LLM Integration
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
